+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.0.6] - 2025-08-13
+### Fixed
+- This version only contains fixes for the Web version..
+
+## [0.0.5] - 2025-08-13
+### Added
+- Environment parameter for `generateNonce` and platform version getters for Android and iOS.
+- Accept.js readiness check.
+### Changed
+- Bumped plugin version and updated dependencies.
+- Refreshed WDePOS SDK support.
+### Fixeds
+- Ensured Accept.js context is validated before dispatching data.
+
 ## [0.0.4] - 2025-08-12
 ### Added
 - Support for WDePOS on iOS and Android.